Implement bit_qtree.{c,h}pp
Initial attempt at storing a quadtree for the purposes of a baseline to
compare against. However, is innapropriate, as is not itself queryable.
Will need to write another class to handle this.

On the upside, bit_qtree is a nice way to *store* quadtrees, and build
other representations of quadtrees from. So it will probably feature
heavily herein in the strategy of constructing other forms of trees.