Extend sdsl/k2_tree as k2_range.hpp
In the style of, and with thanks to, Francisco Montoto (GitHub: fmontoto)

However, do note that this will only function with an edited version of
sdsl -- namely, the private attributes and methods in sdsl::k2_tree must
be made public, or protected. A patch will be submitted to
simongog/sdsl-lite in some point in the future.